+++ Block/System/Config/Webhook.php
@@ -1,0 +1,78 @@
+<?php
+
+namespace Radarsofthouse\Reepay\Block\System\Config;
+
+use Magento\Backend\Block\Template\Context;
+use Magento\Config\Block\System\Config\Form\Field;
+use Magento\Framework\Data\Form\Element\AbstractElement;
+
+class Webhook extends Field
+{
+    /**
+     * @var string
+     */
+    protected $_template = 'Radarsofthouse_Reepay::system/config/webhook.phtml';
+
+    /**
+     * @param Context $context
+     * @param array $data
+     */
+    public function __construct(
+        Context $context,
+        array $data = []
+    ) {
+        parent::__construct($context, $data);
+    }
+
+    /**
+     * Remove scope label
+     *
+     * @param  AbstractElement $element
+     * @return string
+     */
+    public function render(AbstractElement $element)
+    {
+        $element->unsScope()->unsCanUseWebsiteValue()->unsCanUseDefaultValue();
+        return parent::render($element);
+    }
+
+    /**
+     * Return element html
+     *
+     * @param  AbstractElement $element
+     * @return string
+     */
+    protected function _getElementHtml(AbstractElement $element)
+    {
+        return $this->_toHtml();
+    }
+
+    /**
+     * Return ajax url for collect button
+     *
+     * @return string
+     */
+    public function getAjaxUrl()
+    {
+        return $this->getUrl('radarsofthouse_reepay/system_config/webhook');
+    }
+
+    /**
+     * Generate collect button html
+     *
+     * @return string
+     */
+    public function getButtonHtml()
+    {
+        $button = $this->getLayout()->createBlock(
+            'Magento\Backend\Block\Widget\Button'
+        )->setData(
+            [
+                'id' => 'webhook_button',
+                'label' => __('Update Webhook'),
+            ]
+        );
+
+        return $button->toHtml();
+    }
+}